hashmap: randomize enumeration order
As outlined in [0], the enumeration of a hashmap may leak
information about the key used for hashing (possibly negating
the benefit of having such a key to begin with).

This implements the same logic as in Perl:

"In order to prevent this information leakage we randomize the
element order returned by the keys() and each() functions on the
hash. We do this by adding a mask to each hash, and when an
insert into the hash occurs we modify the mask in a pseudo-random
way. During traversal we iterate from 0 to the k-th bucket and
then XOR the iteration value with the mask. The result is that
every time a new key is added to the hash the order of keys will
change more or less completely. This means that the "natural" key
order of the hash exposes almost no useful data to an attacker.
Seeing one key in front of another does not tell you anything
about which bucket the key was stored in."

We get a new random seed every time the hashmap is resized, and we
update this pseudorandomly on every add().

[0]: <http://blog.booking.com/hardening-perls-hash-function.html>