sd-dhcp-lease: add sd_dhcp_lease_{from,get}_raw()
This API allows a lease to be created from the raw binary blob as
received from the network, together when the timestamp the request
for the lease was sent.

The new public API is used internally from sd-dhcp-client, and it
replaces the old internal dhcp_lease_save() call in networkd to
save the lease to disk. dhcp_lease_{load,save}() are now redundant
and are removed.